Added login to admin with sessions

+++ admin/delete.php
@@ -1,12 +1,13 @@
 <?php
 if(isset($_GET['delete'])) {
 
-	$deleteUser = User::deleteRecord($_GET['delete']);
+	if($session->is_signed_in() && $_SESSION['role'] === 'Admin') {
 
-	if($deleteUser) { header("Location: users.php"); }
+		$deleteUser = User::deleteRecord($_GET['delete']);
+
+		if($deleteUser) { header("Location: users.php"); }
+	}
 }
-
-
 ?>
 <!-- Modal -->
 <div id="deleteModal" class="modal fade" role="dialog">
